+++ src/models/ledger/feeSettings.ts
@@ -1,4 +1,4 @@
-import { BaseLedgerEntry } from "./base_ledger_entry";
+import { BaseLedgerEntry } from "./baseLedgerEntry";
 
 export interface FeeSettings extends BaseLedgerEntry {
     LedgerEntryType: 'FeeSettings'
