+++ src/models/ledger/index.ts
@@ -1,17 +1,17 @@
-import { Ticket } from "../../transaction/ticket";
-import { AccountRoot } from "./account_root";
+import { AccountRoot } from "./accountRoot";
 import { Amendments } from "./amendments";
 import { Check } from "./check";
-import { DepositPreauth } from "./deposit_preauth";
-import { DirectoryNode } from "./directory_node";
+import { DepositPreauth } from "./depositPreauth";
+import { DirectoryNode } from "./directoryNode";
 import { Escrow } from "./escrow";
-import { FeeSettings } from "./fee_settings";
-import { LedgerHashes } from "./ledger_hashes";
-import { NegativeUNL } from "./negative_unl";
+import { FeeSettings } from "./feeSettings";
+import { LedgerHashes } from "./ledgerHashes";
+import { NegativeUNL } from "./negativeUNL";
 import { Offer } from "./offer";
-import { PayChannel } from "./pay_channel";
-import { RippleState } from "./ripple_state";
-import { SignerList } from "./signer_list";
+import { PayChannel } from "./payChannel";
+import { RippleState } from "./rippleState";
+import { SignerList } from "./signerList";
+import { Ticket } from "./ticket";
 
 export type LedgerEntry = AccountRoot 
                         | Amendments
