+++ src/models/ledger/negativeUNL.ts
@@ -1,4 +1,4 @@
-import { BaseLedgerEntry } from "./base_ledger_entry";
+import { BaseLedgerEntry } from "./baseLedgerEntry";
 
 interface DisabledValidator {
     FirstLedgerSequence: number
